Fixed a bug in the restate deduction rule
The bug was discovered by Gregor Feierabend. When restating, SOFiA was allowed to rename a variable to a context variable. It was detecting the error in the history but was still performing the substituting substitution. The problem was solved by adding "else:" on Line 302. The error message on Line 131 was also updated, removing "context", since SOFiA will not allow renaming a variable with a variable that appears in the statement being restated.