Flexible assumptions
In a given proof block, previously only the first line was the assumption line. Now any line can be an assumption line. Such assumption lines can be added by the ".aa" command. Inclusion of this feature is inspired by a suggestion of Gregor Feierabend. This command only works inside an existing proof block created by the ".a" command.